# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "A custom PaaS platform for automated application deployment.\nDeploy applications from Git repositories to Kubernetes with zero-configuration.",
        "title": "App Deployer API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/alvesdmateus/app-deployer"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/api/v1",
    "paths": {
        "/analyze": {
            "post": {
                "description": "Analyze source code at a given path to detect language, framework, and dependencies",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analyzer"
                ],
                "summary": "Analyze source code",
                "parameters": [
                    {
                        "description": "Path to analyze",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.AnalyzeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/analyzer.AnalysisResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/analyze/upload": {
            "post": {
                "description": "Upload source code as a file and analyze it to detect language, framework, and dependencies",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "analyzer"
                ],
                "summary": "Upload and analyze source code",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Source code archive (zip, tar.gz)",
                        "name": "source",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/analyzer.AnalysisResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/deployments/{deployment_id}/builds/latest": {
            "get": {
                "description": "Returns the latest build for a specific deployment",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "builds"
                ],
                "summary": "Get latest build",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Deployment ID",
                        "name": "deployment_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.BuildResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/api.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "analyzer.AnalysisResult": {
            "type": "object",
            "properties": {
                "build_command": {
                    "type": "string"
                },
                "build_tool": {
                    "$ref": "#/definitions/analyzer.BuildTool"
                },
                "confidence": {
                    "type": "number"
                },
                "dependencies": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "dev_dependencies": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "files": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "framework": {
                    "$ref": "#/definitions/analyzer.Framework"
                },
                "has_dockerfile": {
                    "type": "boolean"
                },
                "language": {
                    "$ref": "#/definitions/analyzer.Language"
                },
                "port": {
                    "type": "integer"
                },
                "runtime": {
                    "type": "string"
                },
                "start_command": {
                    "type": "string"
                }
            }
        },
        "analyzer.BuildTool": {
            "type": "string",
            "enum": [
                "npm",
                "yarn",
                "pnpm",
                "go",
                "pip",
                "poetry",
                "maven",
                "gradle",
                "cargo",
                "unknown"
            ],
            "x-enum-varnames": [
                "BuildToolNPM",
                "BuildToolYarn",
                "BuildToolPNPM",
                "BuildToolGo",
                "BuildToolPip",
                "BuildToolPoetry",
                "BuildToolMaven",
                "BuildToolGradle",
                "BuildToolCargo",
                "BuildToolUnknown"
            ]
        },
        "analyzer.Framework": {
            "type": "string",
            "enum": [
                "gin",
                "echo",
                "chi",
                "fiber",
                "express",
                "nestjs",
                "nextjs",
                "koa",
                "fastify",
                "flask",
                "django",
                "fastapi",
                "springboot",
                "quarkus",
                "unknown"
            ],
            "x-enum-varnames": [
                "FrameworkGin",
                "FrameworkEcho",
                "FrameworkChi",
                "FrameworkFiber",
                "FrameworkExpress",
                "FrameworkNestJS",
                "FrameworkNextJS",
                "FrameworkKoa",
                "FrameworkFastify",
                "FrameworkFlask",
                "FrameworkDjango",
                "FrameworkFastAPI",
                "FrameworkSpringBoot",
                "FrameworkQuarkus",
                "FrameworkUnknown"
            ]
        },
        "analyzer.Language": {
            "type": "string",
            "enum": [
                "go",
                "nodejs",
                "python",
                "java",
                "rust",
                "ruby",
                "php",
                "dotnet",
                "unknown"
            ],
            "x-enum-varnames": [
                "LanguageGo",
                "LanguageNodeJS",
                "LanguagePython",
                "LanguageJava",
                "LanguageRust",
                "LanguageRuby",
                "LanguagePHP",
                "LanguageDotNet",
                "LanguageUnknown"
            ]
        },
        "api.AnalyzeRequest": {
            "type": "object",
            "properties": {
                "path": {
                    "type": "string"
                }
            }
        },
...
        "api.BuildResponse": {
            "type": "object",
            "properties": {
                "build_log": {
                    "type": "string"
                },
                "completed_at": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "deployment_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "image_tag": {
                    "type": "string"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "api.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        }
    }
}